rename: hma_transfer_config -> ssm_fa_transfer_config, mamba_conv_transfer_utils -> ssm_conv_transfer_utils
