+++ vllm/distributed/kv_transfer/kv_connector/v1/nixl_connector.py
@@ -41,23 +41,23 @@ from vllm.distributed.kv_transfer.kv_connector.v1.base import (
     KVConnectorRole,
     SupportsHMA,
 )
-from vllm.distributed.kv_transfer.kv_connector.v1.hma_transfer_config import (
-    HeteroTPTransferConfig,
-    create_transfer_config,
-    fa_divisor,
-    should_skip_fa,
-    uses_split_handles,
-)
-from vllm.distributed.kv_transfer.kv_connector.v1.mamba_conv_transfer_utils import (
-    MambaConvSplitInfo,
-    compute_mamba_phys_ratio,
-    derive_mamba_conv_split,
-)
 from vllm.distributed.kv_transfer.kv_connector.v1.metrics import (
     KVConnectorPromMetrics,
     KVConnectorStats,
     PromMetric,
     PromMetricT,
+)
+from vllm.distributed.kv_transfer.kv_connector.v1.ssm_conv_transfer_utils import (
+    MambaConvSplitInfo,
+    compute_mamba_phys_ratio,
+    derive_mamba_conv_split,
+)
+from vllm.distributed.kv_transfer.kv_connector.v1.ssm_fa_transfer_config import (
+    HeteroTPTransferConfig,
+    create_transfer_config,
+    fa_divisor,
+    should_skip_fa,
+    uses_split_handles,
 )
 from vllm.distributed.parallel_state import (
     get_tensor_model_parallel_rank,
